relation controller bugs fixed

+++ controllers/user-movie-controllers.js
@@ -3,7 +3,26 @@ const UserMovie = require("../models/user-movie");
 const addUserMovieRelation = async (req, res, next) => {
     const { movie, relation_type, like } = req.body;
 
+    let relation;
+    try {
+        relation = await UserMovie.findOne({
+            user: req.auth.sub,
+            movie,
+            relation_type,
+        });
+    } catch (err) {
+        const error = new Error("Fetching relation failed.");
+        error.message = 500;
+        return next(error);
+    }
+
     if (!like) {
+        if (!relation) {
+            const error = new Error("Relation doesn't exist");
+            error.status = 404;
+            return next(error);
+        }
+
         try {
             await UserMovie.deleteOne({
                 user: req.auth.sub,
@@ -19,19 +38,6 @@ const addUserMovieRelation = async (req, res, next) => {
         return res.json({
             message: "relation deleted successfully!",
         });
-    }
-
-    let relation;
-    try {
-        relation = await UserMovie.find({
-            user: req.auth.sub,
-            movie,
-            relation_type,
-        });
-    } catch (err) {
-        const error = new Error("Fetching relation failed.");
-        error.message = 500;
-        return next(error);
     }
 
     if (relation) {
